backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
psycopg2-binary>=2.9,<3.0
django-cors-headers>=4.3,<5.0
djangorestframework-simplejwt>=5.3,<6.0
pandas>=2.1,<3.0
requests>=2.31,<3.0
gunicorn>=21.2,<23.0
python-dotenv>=1.0,<2.0
anthropic>=0.39,<1.0
yfinance>=0.2,<1.0
